KULRICE-4437 - committing first merge of 1.0.3 into 1.1.0 - relevant 1.0.3 revision is 14215

+++ impl/src/main/java/org/kuali/rice/kew/actionlist/web/ActionListDecorator.java
@@ -47,7 +47,7 @@ public class ActionListDecorator extends TableDecorator {
                 Integer index = actionItem.getActionItemIndex();
                 Integer frameHeight = new Integer(290);
                 try {
-                    if (displayParameters.getFrameHeight() != null && displayParameters.getFrameHeight().intValue() > 0) {
+                    if (displayParameters.getFrameHeight() != null && displayParameters.getFrameHeight() > 0) {
                         frameHeight = displayParameters.getFrameHeight();
                     }
                 } catch (Exception ex) {
